requirements.txt
# requirements.txt — CTI-RL-MTD Honey Drone Testbed
#
# [설치]  pip install -r requirements.txt
# [환경]  Python 3.11+ / WSL2 Ubuntu 22.04
#
# ── Core ──────────────────────────────────────────────────────────
pymavlink==2.4.41           # MAVLink 패킷 파싱 (DroneKit 의존성 없이 사용)
python-dotenv==1.0.1        # .env 로드 (constants.py)

# ── CTI / STIX ────────────────────────────────────────────────────
stix2==3.0.1                # STIX 2.1 번들 생성 및 직렬화
taxii2-client==2.3.0        # TAXII 서버 연동 (선택)

# ── API ───────────────────────────────────────────────────────────
fastapi==0.115.0            # CTI Ingest API
uvicorn[standard]==0.30.1   # ASGI 서버
pydantic==2.7.0             # Request/Response 모델 검증

# ── RL ────────────────────────────────────────────────────────────
stable-baselines3==2.3.2    # PPO 구현 (기존 TIFS 코드베이스 호환)
gymnasium==0.29.1           # 환경 인터페이스 (OpenAI Gym 후속)
torch==2.3.1                # SB3 백엔드

# ── 데이터 처리 ───────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2

# ── 로깅 / 관측성 ─────────────────────────────────────────────────
structlog==24.2.0           # JSON 구조화 로깅 (논문 재현용)

# ── 통계 / 시각화 ─────────────────────────────────────────────────
scipy==1.13.1               # Wilcoxon rank-sum 검정 (Phase 4.4)
matplotlib==3.9.0           # 논문 figure 생성 (Phase 4.5)
seaborn==0.13.2             # 보조 시각화

# ── 테스트 ────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# ── 코드 품질 ─────────────────────────────────────────────────────
black==24.4.2
isort==5.13.2
flake8==7.0.0
mypy==1.10.0
